requirements.txt
black==24.10.0
click==8.1.7
mypy-extensions==1.0.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
